restaurant_recommender7_labeled/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
pandas==2.0.3
joblib==1.3.2
numpy==1.24.3